xwayland: Allow pointer warps even when cursor is visible.
This fixes XWarpPointer being no-op on XWayland even when the cursor is
within the focused window surface. This is something that works with X
server, and Wine expects that to implement Windows SetCursorPos for many
games.

Using zwp_locked_pointer_v1 should already make sure we're not warping
pointer when outside of the focused window surface.