xwayland: Check for pointer caps on lock / warp / confine
The pointer locking protocol relies on the wl_pointer, if the pointer
capability was lost, the wl_pointer is NULL.

In that case, trying to use the pointer locking, either to emulate a
lock or a pointer warp, will lead to Wayland protocol error which is
fatal to Xwayland.

To avoid that issue, make sure we (still) have a wl_pointer available
before trying to lock or warp the pointer.

If not, record the last window and coordinate from the request and
resubmit the request as soon as the wl_pointer is restored.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Suggested-by: Michel Dänzer <michel@daenzer.net>
Closes: https://gitlab.freedesktop.org/xorg/xserver/-/issues/1829